+++ AsyncLoggers/src/BepInExListeners/SqliteLogger.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
+using AsyncLoggers.Config;
 using BepInEx;
 using SQLite;